업데이트

+++ src/main/java/com/github/riset_backend/manageCompany/service/MailService.java
@@ -1,5 +1,7 @@
 package com.github.riset_backend.manageCompany.service;
 
+import ch.qos.logback.core.spi.ErrorCodes;
+import com.github.riset_backend.global.config.exception.ErrorCode;
 import com.github.riset_backend.manageCompany.entity.Email;
 import com.github.riset_backend.manageCompany.repository.EmailRepository;
 import com.github.riset_backend.global.config.auth.JwtTokenProvider;
@@ -40,28 +42,30 @@ public class MailService {
     private String sendEmail;
 
     // 이메일 발송
-    public void answerMail(String email, CustomUserDetails user) {
+    public String answerMail(String email, CustomUserDetails user) {
 
         //유저 확인
-        Optional<Employee> userName = Optional.ofNullable(employeeRepository.findByEmployeeId(user.getUsername()).orElseThrow(() -> new BusinessException(NOT_USER, "유저가 없어요")));
+        Optional<Employee> userName = Optional.ofNullable(employeeRepository.findByEmployeeId(user.getUsername()).orElseThrow(
+                () -> new BusinessException(NOT_USER, "없는 유저 입니다.")));
 
-        //회사 확인
-        if (userName.get().getCompany().getCompanyNo() != null) {
+
+        //회사 확인, ADMIN 권한확인
+        if (user.getEmployee().getRoles().name().equals(Role.ROLE_ADMIN.name()) && userName.get().getCompany().getCompanyNo() != null) {
             Long companyNo = userName.get().getCompany().getCompanyNo();
-            try {
-                String code = RandomCodeGenerator.generateCode() + companyNo;
-                // 이메일 전송
-                sendVerificationEmail(email, code);
-                Email emailEntity = Email.builder()
-                        .code(code)
-                        .employeeId(userName.get().getEmployeeNo())
-                        .build();
 
-                emailRepository.save(emailEntity);//db 저장
-            } catch (MailException e) {
-                throw new RuntimeException(e);
-            }
+            String code = RandomCodeGenerator.generateCode() + companyNo;
+            // 이메일 전송
+            sendVerificationEmail(email, code);
+            Email emailEntity = Email.builder()
+                    .code(code)
+                    .employeeId(userName.get().getEmployeeNo())
+                    .build();
+
+            emailRepository.save(emailEntity);//db 저장
+        } else {
+            return "발급 실패";
         }
+        return email + "에 이메일을 발송하였습니다";
     }
 
 
@@ -69,40 +73,45 @@ public class MailService {
     public String CompanyCode(String code, String token, CustomUserDetails user) {
 
         String jwt = token.substring(7);
-        //직원 id
-        try {
-            // 이메일 리포지토리에서 코드를 사용하여 직원 ID를 가져옴
-            Email emailUser = emailRepository.findByCode(code)
-                    .orElseThrow(() -> new BusinessException(NOT_USER, "이메일이 코드와 연관되어 있지 않습니다."));
 
-            // 코드에서 숫자만 추출하여 회사 ID로 사용
-            Long numbersOnly = Long.valueOf(code.replaceAll("[^0-9]", ""));
+        if(!user.getEmployee().getRoles().name().equals(Role.ROLE_ADMIN.name())) {
+            try {
+                // 이메일 리포지토리에서 코드를 사용하여 직원 ID를 가져옴
+                Email emailUser = emailRepository.findByCode(code)
+                        .orElseThrow(() -> new BusinessException(NOT_USER, "이메일이 코드와 연관되어 있지 않습니다."));
 
-            // 회사를 찾음
-            Company companyNo = companyRepository.findById(numbersOnly)
-                    .orElseThrow(() -> new BusinessException(NOT_USER, "해당 회사를 찾을 수 없습니다."));
+                // 코드에서 숫자만 추출하여 회사 ID로 사용
+                Long numbersOnly = Long.valueOf(code.replaceAll("[^0-9]", ""));
 
-            // 직원을 찾음
-            Employee employee = employeeRepository.findByEmployeeNo(user.getEmployee().getEmployeeNo())
-                    .orElseThrow(() -> new BusinessException(NOT_USER, "해당 직원을 찾을 수 없습니다."));
+                // 회사를 찾음
+                Company companyNo = companyRepository.findById(numbersOnly)
+                        .orElseThrow(() -> new BusinessException(NOT_USER, "해당 회사를 찾을 수 없습니다."));
 
-            // 직원의 회사를 업데이트
-            employee.setCompany(companyNo);
-            jwtTokenProvider.setRole(jwt, Role.ROLE_EMPLOYEE.name());
-            employee.setRoles(Role.ROLE_EMPLOYEE);
+                // 직원을 찾음
+                Employee employee = employeeRepository.findByEmployeeNo(user.getEmployee().getEmployeeNo())
+                        .orElseThrow(() -> new BusinessException(NOT_USER, "해당 직원을 찾을 수 없습니다."));
+
+                // 직원의 회사를 업데이트
+                employee.setCompany(companyNo);
+                jwtTokenProvider.setRole(jwt, Role.ROLE_EMPLOYEE.name());
+                employee.setRoles(Role.ROLE_EMPLOYEE);
 
 
-            employeeRepository.save(employee);
-            emailRepository.delete(emailUser);
-            return "인증되었습니다";
+                employeeRepository.save(employee);
+                emailRepository.delete(emailUser);
+                return companyNo.getCompanyName() + "회사에 등록되었습니다";
 
-        } catch (NumberFormatException e) {
-            // 숫자로 변환할 수 없는 코드일 때 발생하는 예외 처리
-            throw new BusinessException(NOT_USER, "잘못된 코드 형식입니다.");
-        } catch (Exception e) {
-            // 기타 예외 처리
-            throw new BusinessException(NOT_USER, "직원 정보를 저장하는 동안 오류가 발생했습니다.", e);
+            } catch (NumberFormatException e) {
+                // 숫자로 변환할 수 없는 코드일 때 발생하는 예외 처리
+                throw new BusinessException(NOT_USER, "잘못된 코드 형식입니다.");
+            } catch (Exception e) {
+                // 기타 예외 처리
+                throw new BusinessException(NOT_USER, "직원 정보를 저장하는 동안 오류가 발생했습니다.", e);
+            }
         }
+        //직원 id
+
+        return  "등록 실패";
     }
